pyproject.toml
[project]
name = "masterthesis-llm-testing"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "openai>=1.40",   # openai als zugriffsmethode auf llms
  "pydantic>=2.7",  # json handling
  "pyyaml>=6.0",    # anwendungsfälle aus yaml
  "python-dotenv>=1.0"  # .env handling 
]
